修改第三方登入資料不齊全的處理，新增檔案將passport的設定分出來
