+++ app.js
@@ -6,6 +6,7 @@ var cors = require('cors');
 const swaggerUI = require('swagger-ui-express');
 const swaggerFilePath = `./swagger-output_${process.env.NODE_ENV}.json`;
 const swaggerFile = require(swaggerFilePath);
+const passport = require('passport');
 
 /* router ------------------------------------------------------------------- */
 var usersRouter = require('./routes/users');
@@ -33,6 +34,9 @@ app.use('/tp-auth', tpAuthRouter);
 app.use('/', indexRouter);
 app.use('/api-doc', swaggerUI.serve, swaggerUI.setup(swaggerFile));
 /* /express 設定 --------------------------------------------------------------- */
+/* /passport 設定 --------------------------------------------------------------- */
+require('./config/passport')(passport);
+/* /passport 設定 --------------------------------------------------------------- */
 
 // 404 錯誤
 app.use(function (req, res, next) {
