+++ config/passport.js
@@ -1,0 +1,31 @@
+const GoogleStrategy = require('passport-google-oauth20').Strategy;
+const FacebookStrategy = require('passport-facebook').Strategy;
+
+module.exports = (passport) => {
+  passport.use(
+    new GoogleStrategy(
+      {
+        clientID: process.env.GOOGLE_CLIENT_ID,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+        callbackURL: process.env.GOOGLE_CALLBACK_URL,
+        profileFields: ['email', 'displayName'],
+      },
+      (accessToken, refreshToken, profile, cb) => {
+        return cb(null, profile);
+      }
+    )
+  );
+  passport.use(
+    new FacebookStrategy(
+      {
+        clientID: process.env.FACEBOOK_CLIENT_ID,
+        clientSecret: process.env.FACEBOOK_CLIENT_SECRET,
+        callbackURL: process.env.FACEBOOK_CALLBACK_URL,
+        profileFields: ['id', 'displayName', 'photos', 'email'],
+      },
+      function (accessToken, refreshToken, profile, cb) {
+        return cb(null, profile);
+      }
+    )
+  );
+};
