+++ controllers/tpAuth.js
@@ -5,10 +5,19 @@ const handleError = require('../handStates/handleError');
 const auth = {
   googleCallback: handleError(async (req, res) => {
     const user = await findOrCreateUser('google', req.user._json);
+    if (user === null) {
+      res.redirect(`${process.env.FRONTEND_MEMBER_URL}/login`);
+    }
+    // res.status(200).json(user);
     generateRedirectJWT(user, res);
   }),
   facebookCallback: handleError(async (req, res) => {
+    // console.log('req.user._json', req.user._json);
     const user = await findOrCreateUser('facebook', req.user._json);
+    if (user === null) {
+      res.redirect(`${process.env.FRONTEND_MEMBER_URL}/login`);
+    }
+    // res.status(200).json(user);
     generateRedirectJWT(user, res);
   }),
 };
